+++ src/core/element_container/GeneratorContainer.hpp
@@ -73,20 +73,20 @@ class LS2G_API GeneratorContainer final: public OneSideContainer_PQ, public Iter
         virtual ~GeneratorContainer() noexcept = default;
         
         // TODO add pmin and pmax here !
-        void init(const RealVect & generators_p,
-                  const RealVect & generators_v,
-                  const RealVect & generators_min_q,
-                  const RealVect & generators_max_q,
-                  const Eigen::VectorXi & generators_bus_id
+        void init(const Eigen::Ref<const RealVect> & generators_p,
+                  const Eigen::Ref<const RealVect> & generators_v,
+                  const Eigen::Ref<const RealVect> & generators_min_q,
+                  const Eigen::Ref<const RealVect> & generators_max_q,
+                  const Eigen::Ref<const Eigen::VectorXi> & generators_bus_id
                   );
-              
-        void init_full(const RealVect & generators_p,
-                       const RealVect & generators_v,
-                       const RealVect & generators_q,
+
+        void init_full(const Eigen::Ref<const RealVect> & generators_p,
+                       const Eigen::Ref<const RealVect> & generators_v,
+                       const Eigen::Ref<const RealVect> & generators_q,
                        const std::vector<bool> & voltage_regulator_on,
-                       const RealVect & generators_min_q,
-                       const RealVect & generators_max_q,
-                       const Eigen::VectorXi & generators_bus_id
+                       const Eigen::Ref<const RealVect> & generators_min_q,
+                       const Eigen::Ref<const RealVect> & generators_max_q,
+                       const Eigen::Ref<const Eigen::VectorXi> & generators_bus_id
                        );
                    
         // pickle
@@ -156,7 +156,7 @@ class LS2G_API GeneratorContainer final: public OneSideContainer_PQ, public Iter
             const Eigen::Ref<const RealVect> & /*Vm*/,
             const Eigen::Ref<const CplxVect> & /*V*/,
             const SolverBusIdVect & /*id_grid_to_solver*/,
-            const RealVect & /*bus_vn_kv*/,
+            const Eigen::Ref<const RealVect> & /*bus_vn_kv*/,
             real_type /*sn_mva*/,
             bool ac) override {
               set_osc_pq_res_p();