+++ src/core/element_container/SvcContainer.cpp
@@ -15,13 +15,13 @@
 namespace ls2g {
 
 void SvcContainer::init(const std::vector<int> & regulation_mode,
-                        const RealVect & target_vm_pu,
-                        const RealVect & q_setpoint_mvar,
-                        const RealVect & slope_pu,
-                        const RealVect & b_min,
-                        const RealVect & b_max,
-                        const Eigen::VectorXi & regulated_bus_id,
-                        const Eigen::VectorXi & bus_id)
+                        const Eigen::Ref<const RealVect> & target_vm_pu,
+                        const Eigen::Ref<const RealVect> & q_setpoint_mvar,
+                        const Eigen::Ref<const RealVect> & slope_pu,
+                        const Eigen::Ref<const RealVect> & b_min,
+                        const Eigen::Ref<const RealVect> & b_max,
+                        const Eigen::Ref<const Eigen::VectorXi> & regulated_bus_id,
+                        const Eigen::Ref<const Eigen::VectorXi> & bus_id)
 {
     const int size = static_cast<int>(bus_id.size());
     // an SVC has NO active power
@@ -155,7 +155,7 @@ void SvcContainer::_compute_results(
     const Eigen::Ref<const RealVect> & /*Vm*/,
     const Eigen::Ref<const CplxVect> & /*V*/,
     const SolverBusIdVect & /*id_grid_to_solver*/,
-    const RealVect & /*bus_vn_kv*/,
+    const Eigen::Ref<const RealVect> & /*bus_vn_kv*/,
     real_type /*sn_mva*/,
     bool ac)
 {